requirements.txt
flask
Gunicorn
sqlalchemy
pymysql
python-dotenv